Fix dns/firewall code to work with new structure
