+++ talpid-core/Cargo.toml
@@ -43,6 +43,7 @@ rtnetlink = { git = "https://github.com/mullvad/netlink", branch = "best-effort-
 nftnl = { git = "https://github.com/mullvad/nftnl-rs", rev = "f0b1492fd2fd1f737dbffd047c9c60c300e6f7d6", features = ["nftnl-1-1-0"] }
 mnl = { git = "https://github.com/mullvad/mnl-rs", rev = "f0d19501b9b85be9a1ffaec8317a378bcbdf4fa6", features = ["mnl-1-0-4"] }
 which = "2.0"
+err-derive = "0.1.5"
 
 [target.'cfg(target_os = "macos")'.dependencies]
 pfctl = "0.2"
