+++ talpid-core/src/dns/mod.rs
@@ -1,3 +1,5 @@
+use std::{net::IpAddr, path::Path};
+
 #[cfg(target_os = "macos")]
 #[path = "macos.rs"]
 mod imp;
@@ -10,6 +12,47 @@ mod imp;
 #[path = "windows.rs"]
 mod imp;
 
-pub use self::imp::{DnsError, Error};
+pub use self::imp::Error;
 
-pub use self::imp::{DnsMonitor, Error as DnsError};
+/// Sets and monitors system DNS settings. Makes sure the desired DNS servers are being used.
+pub struct DnsMonitor {
+    inner: imp::DnsMonitor,
+}
+
+impl DnsMonitor {
+    /// Returns a new `DnsMonitor` that can set and monitor the system DNS.
+    pub fn new(cache_dir: impl AsRef<Path>) -> Result<Self, Error> {
+        Ok(DnsMonitor {
+            inner: imp::DnsMonitor::new(cache_dir)?,
+        })
+    }
+
+    /// Set DNS to the given servers. And start monitoring the system for changes.
+    pub fn set(&mut self, interface: &str, servers: &[IpAddr]) -> Result<(), Error> {
+        log::info!(
+            "Setting DNS servers to {}",
+            servers
+                .iter()
+                .map(|ip| ip.to_string())
+                .collect::<Vec<String>>()
+                .join(", ")
+        );
+        self.inner.set(interface, servers)
+    }
+
+    /// Reset system DNS settings to what it was before being set by this instance.
+    pub fn reset(&mut self) -> Result<(), Error> {
+        log::info!("Resetting DNS");
+        self.inner.reset()
+    }
+}
+
+trait DnsMonitorT: Sized {
+    type Error: ::std::error::Error;
+
+    fn new(cache_dir: impl AsRef<Path>) -> Result<Self, Self::Error>;
+
+    fn set(&mut self, interface: &str, servers: &[IpAddr]) -> Result<(), Self::Error>;
+
+    fn reset(&mut self) -> Result<(), Self::Error>;
+}
