+++ talpid-core/src/lib.rs
@@ -40,7 +40,14 @@ pub mod mpsc;
 /// Abstractions over operating system network security settings.
 pub mod security;
 
+/// Abstractions over operating system DNS settings.
+pub mod dns;
+
 /// State machine to handle tunnel configuration.
 pub mod tunnel_state_machine;
 
 mod mktemp;
+
+/// Misc utilities for the Linux platform.
+#[cfg(target_os = "linux")]
+mod linux;
